- remove implicit conversions and comparisons between different types - compiles now clean with  -Wconversion -Wsign-conversion -Wfloat-conversion -Wsign-compare

+++ nscgi/nscgi.c
@@ -1144,7 +1144,7 @@ CgiCopy(Cgi *cgiPtr, Ns_Conn *conn)
                 ++value;
             }
             if (STRIEQ(ds.string, "status")) {
-                httpstatus = strtol(value, NULL, 10);
+                httpstatus = (int)strtol(value, NULL, 10);
             } else if (STRIEQ(ds.string, "location")) {
                 httpstatus = 302;
                 if (*value == '/') {
