+++ nsd/returnresp.c
@@ -91,7 +91,7 @@ ConfigServerRedirects(const char *server)
 
         key = Ns_SetKey(set, i);
         map = Ns_SetValue(set, i);
-        statusCode = strtol(key, NULL, 10);
+        statusCode = (int)strtol(key, NULL, 10);
         if (statusCode <= 0 || *map == '\0') {
             Ns_Log(Error, "redirects[%s]: invalid redirect '%s=%s'",
                    server, key, map);
